requirements.txt
# Requirements pour le projet olympique
pandas>=2.0.0
mysql-connector-python>=8.0.0
openpyxl>=3.1.0
lxml>=4.9.0
html5lib>=1.1
scikit-learn>=1.2.0
seaborn>=0.12.0
matplotlib>=3.7.0
pyyaml>=6.0
joblib>=1.3.0